csv_upload_back/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
openapi-codec==1.3.2
packaging==23.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
